% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggxts.R
\name{ggxts}
\alias{ggxts}
\title{ggxts}
\usage{
ggxts(x, digits = 2)
}
\arguments{
\item{x}{the xts object}

\item{digits}{digits of ene values}
}
\value{
ggplot object
}
\description{
plot xts with end values via ggplot
}
