% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxplot.R
\name{mxplot.gg}
\alias{mxplot.gg}
\title{mxplot.gg}
\usage{
\method{mxplot}{gg}(..., use_one_x_axis = T, theme = theme_bw())
}
\arguments{
\item{...}{ggplot objects}

\item{use_one_x_axis}{if true, only the last plot show x-axis}

\item{theme}{the theme}
}
\description{
plot multiple xts objects by row
}
