% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxplot.R
\name{mxplot.xts}
\alias{mxplot.xts}
\title{mxplot.xts}
\usage{
\method{mxplot}{xts}(..., use_one_x_axis = T, theme = theme_bw(),
  titles = NULL)
}
\arguments{
\item{...}{xts objects}

\item{use_one_x_axis}{if true, only the last plot show x-axis}

\item{theme}{the theme}

\item{titles}{the vector of titles}
}
\description{
plot multiple xts objects by row
}
