% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxplot.R
\name{mxplotList}
\alias{mxplotList}
\title{mxplotList}
\usage{
mxplotList(plot_list, use_one_x_axis = T, theme = theme_bw(),
  titles = NULL)
}
\arguments{
\item{plot_list}{a list of ggplot objects}

\item{use_one_x_axis}{if true then remove the x-axis of plots but the last}

\item{theme}{the theme to apply}

\item{titles}{a vector of titles}
}
\description{
mxplotList
}
