% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxplot.R
\name{mxplot.gg}
\alias{mxplot.gg}
\title{mxplot.gg}
\usage{
\method{mxplot}{gg}(..., use_one_x_axis = T, theme = theme_bw(),
  heights = NULL, end_spacing = 0.1)
}
\arguments{
\item{...}{ggplot objects}

\item{use_one_x_axis}{if true, only the last plot show x-axis}

\item{theme}{the theme}

\item{heights}{the weight of the height of plots, eg. heights = c(2,1)}

\item{end_spacing}{extend the space in the end, as a percentage of the size of xlim().}
}
\description{
plot multiple xts objects by row
}
