% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxplot.R
\name{mxplot.xts}
\alias{mxplot.xts}
\title{mxplot.xts}
\usage{
\method{mxplot}{xts}(..., use_one_x_axis = T, theme = theme_bw(),
  titles = NULL, heights = NULL, size = 0.8, end_spacing = 0.1)
}
\arguments{
\item{...}{xts objects}

\item{use_one_x_axis}{if true, only the last plot show x-axis}

\item{theme}{the theme}

\item{titles}{the vector of titles}

\item{heights}{the weight of the height of plots, eg. heights = c(2,1)}

\item{size}{line width}

\item{end_spacing}{extend the space in the end, as a percentage of the size of xlim().}
}
\description{
plot multiple xts objects by row
}
